requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
websockets==11.0.3
websocket-client==1.6.4
watchdog>=2.3.0
python 3.11.5